+++ backend/src/main/java/de/kaliburg/morefair/statistics/results/RoundStatisticsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface RoundStatisticsRepository extends MongoRepository<RoundStatisticsEntity, Long> {
 
-  Optional<RoundStatisticsEntity> findByRoundId(Long id);
+  Optional<RoundStatisticsEntity> findByRoundId(Integer id);
 }
